chore: add lightweight utility classnames
This utility function comes in handy when we have to add multiple
classnames in a single component. Sometimes we may want to add
classnames based on certain conditions. This utility is best suited for
that very purpose.

One simple example would be

```jsx
const props = ...;
<Input
  id={'name'}
  className={classnames('input-field', {'has-error': !!errors['name']})}
  {...props} />
```
In the example above, Input component will have `has-error` classname as
long as `errors['name']` is truthy.